refactor(egress): introduce ProviderAdapter scaffold (ADR-039 PR-A)
Factor the inline LiteLLM and Portkey call paths out of ai_gateway.py
behind a ProviderAdapter protocol. Zero behaviour change on the wire:
same provider resolution, same error tags, same audit log lines, same
default backend (litellm_embedded), same external API surface
(dispatch, dispatch_stream, GatewayError, GatewayResult,
StreamingDispatch, scrub_secrets).

Files

  mcp_proxy/egress/adapters/__init__.py
    resolve_adapter(backend) -> ProviderAdapter
    Raises GatewayError(501) for unknown backend, including the
    "cullis_native" slot reserved for PR-B/C/D.

  mcp_proxy/egress/adapters/base.py
    ProviderAdapter Protocol + DispatchContext dataclass. ctx carries
    the trusted-identity envelope (agent_id, org_id, trace_id) the
    router reconstructed from the DPoP-bound cert, plus the optional
    http_client used only by the legacy Portkey adapter.

  mcp_proxy/egress/adapters/litellm.py
    LiteLLMAdapter wrapping the previous _call_litellm_embedded and
    _build_litellm_stream bodies verbatim, including the
    _LITELLM_ERROR_MAP and _map_litellm_exception helpers. Lazy import
    of the litellm package retained so deployments that pin a
    non-LiteLLM backend keep booting without the dep.

  mcp_proxy/egress/adapters/portkey.py
    PortkeyAdapter wrapping the previous _call_portkey body verbatim.
    Anthropic-only gate, error scrubbing on upstream 4xx echoes, and
    the CULLIS_FORWARD_HEADERS constant live here now. Streaming on
    Portkey was never wired; stream_chat_completion raises the same
    501 the dispatcher used to.

  mcp_proxy/egress/ai_gateway.py
    Shrunk from 691 to ~230 lines. Keeps the shared dispatch
    primitives (GatewayResult / StreamingDispatch / GatewayError),
    the secret-scrub helpers, and _resolve_provider_creds. dispatch()
    and dispatch_stream() now resolve the adapter for the configured
    backend and delegate.

Rationale
  ADR-039 (imp/adrs/, gitignored) lays out the migration off LiteLLM
  in the AI dispatch critical path. PR-A is the refactor scaffold:
  separate the call-path implementation from the dispatcher so the
  native Anthropic / OpenAI / Ollama adapters can be added in
  subsequent PRs without touching the dispatcher again. The defensive
  pin to litellm>=1.83.7,<2.0 already landed in PR #987 against the
  April-May 2026 CVE cluster (CVE-2026-42208 on CISA KEV,
  CVE-2026-42271 RCE, CVE-2026-35029 RCE) and is unaffected by this
  refactor.

Compatibility
  - MCP_PROXY_AI_GATEWAY_BACKEND values unchanged
    (litellm_embedded default, portkey legacy).
  - Audit row backend tag unchanged ("litellm_embedded" / "portkey").
  - All external imports from ai_gateway still resolve
    (dispatch, dispatch_stream, GatewayError, StreamingDispatch).
  - No external module referenced the moved internals
    (_call_litellm_embedded, _call_portkey, _build_litellm_stream,
    _LITELLM_ERROR_MAP, _map_litellm_exception,
    CULLIS_FORWARD_HEADERS), verified via grep across mcp_proxy,
    tests, cullis_sdk.

Verification
  Import boot of mcp_proxy.egress.{ai_gateway, llm_chat_router,
  anthropic_messages_router} succeeds. resolve_adapter dispatches
  correctly to LiteLLMAdapter and PortkeyAdapter; the unknown
  backend path raises GatewayError(501) with the expected reason
  tag. isinstance(LiteLLMAdapter(), ProviderAdapter) and the
  PortkeyAdapter counterpart both hold (runtime_checkable Protocol).
